Require a "dead_child_handler($childNum)" method.
NOTE: I've noticed that the registered tick handler is sometimes called like 
"extendedClass::sanity_check()" instead of "multiThread::sanity_check()", which 
causes a warning or two.  It's possible that the registered tick handler needs 
to just be a global function, defined at the top of the library file, and it 
can be passed &$this so it can re-use the proper file handles... ugh.

/multiThreadClass.abstract.php:
	* MAIN:::
		-- added "dead_child_handler($childNum)" as an abstract method: each 
		time clean_children() finds a dead child, this method will be called.
	* __construct():
		-- changed the register_ticks() function call to NOT have an argument, 
		since all arguments are ignored for now... 
	* clean_children():
		-- added header.
		-- each time a dead child is found, it calls the new abstract method 
		"dead_child_handler()" so each finished child can be handled properly.
	* child_death():
		-- added header.