Some very basic fixes and changes to begin dealing with IPC (inter-process communication).
NOTE: IPC doesn't work as it should just yet, but I think it's because
all 
processes are sharing the same queue.  Next change will have different
queues 
for each process (including the parent), and the parent will attach
itself to 
all the queues of it's children, and each child will attach itself to
the queue 
of the parent (queue names will be based upon PIDs).

NOTE2:: at present, the main process won't die unless it's killed by the
user 
that started it...

 * MAIN:::
    -- Uses cs_globalFunctions (in the cs-content directory).
    -- main class extends ipc{} for communication.
 * __construct():
    -- argument for register_tick_function() and
unregister_tick_function() stored